+++ HISTORY.md
@@ -1,4 +1,11 @@
 
+2.2.2 / 2017-02-16
+==================
+
+  * Add option to let customers decide how they would like their custom nested traits to be handled
+  * Fix unsubscribe bug
+  * Add special hanlding for Rich Link
+
 2.1.2 / 2017-02-02
 ==================